+++ src/main/scala/com/jcdecaux/datacorp/spark/storage/SparkRepository.scala
@@ -66,7 +66,7 @@ trait SparkRepository[T] extends Repository[T] with CassandraConnector with CSVC
   }
 
   def findByCondition(conditions: Set[Condition])(implicit encoder: Encoder[T]): Dataset[T] = {
-    import com.jcdecaux.datacorp.spark.util.FilterUtils._
+    import com.jcdecaux.datacorp.spark.util.FilterImplicits._
 
     val df = read()
     if (conditions.nonEmpty && !conditions.toSqlRequest.isEmpty) {
@@ -94,7 +94,6 @@ trait SparkRepository[T] extends Repository[T] with CassandraConnector with CSVC
   /**
     *
     * @param data
-    * @param saveMode Only usable for file storage (Parquet and CSV)
     * @param encoder
     * @return
     */